Fixed some issues with invalidness.
Invalid instruction will return an empty string, rather then trying to
return anything. This is need, since we still need to push something to
an asm queue on a disassembler. Maybe it would be a better idea to move,
this logic to the disasm, so that particular backend won't bother about
this.

Removed fresh mcinst, I hope that just cleaning it would be sufficient

Removed my reimplementation for target lookup, after I removed gcc, from
the stack it starts to work.