Fixes several issues in CFG reconstruction.
This several real and one potential bug.

Bug #1 (real)
-------------
A fallthrough was added automatically, when we reach the end of block
without the destination, when the next address starts another block. Now
it is added only if the end of block is implicit, i.e. it wasn't
previously recognized as end.

Bug #2 (real)
-------------
Blocks, that terminates without any destinations were not added to the
set of terminators and thus weren't recognized as block ends and the
second stage of reconstruction.

Bug #3 (real)
-------------
Some instructions, (c.f., #117) are not recognized correctly by the
lifter, but still pretty understandable to our backend, e.g., is_return
evaluates to true.  To get rid of excessive fall-throughs I've added an
explicit check.

Bug #4 (potential)
------------------

possible - if a list of destinations, extracted from instruction
contains explicit fallthrough (that is currently impossible), then
we will recognized it as a jump. Just in case, if someone will change
the behavior of `dests_of_bil` or `dests_of_basic`, I've made check more
explicit.