Propagate BIL transformations though plugins.
This PR provides a new field in a project record: a function that will
transform instruction list to a BIL program. Initially this
transformation includes all optimizations that were requested from the
command line. Plugin, can rebind this optimization.

There is one issue, that should be qualified, the call
`project.bil_of_insns` is statically bound to the `project`, so there is
no open recursion here or any late binding. So all your changes to the
project will not affect this function, until the next plugin, where it
will be rebound to the new state of project.

I'm considering to switch project type to the object type.