Some new optimizations and fixed arm-gnueabi
This is a snapshot of a work-in-progress, to provide recognition of
parameters, structures and locals to arm programs. I'm commiting this
since it is needed for the plugins. Other than fixes to arm gnueabi
implementation it also introduces `Block.dfs` function, that will
return a sequence of block successors or predecessors in a DFS
order (including the block itself). It also introduces several new
arithmetic optimizations, as well as some other optimizations, like
load/store for example.