Add command line arguments for plugins
This PR makes available command line arguments in the plugin.

Only a restricted subset of command line language is allowed, as all
arguments must be non-positional with long name option (i.e., start with
`--`), and values (if any) must be provided with `=` sign,
e.g. `--callstrings-interesting=st.cpy`. Parameters must be prefixed
with the name of plugin to which they should be passed, e.g.,

```
bap-objdump /bin/ls -lstaticstore --staticstore-help
```
A set of parameters is passed to a plugin in a `argv` field of a project
record with removed prefixes, e.g., `--staticstore-help` is transformed
to `--help`;  `--staticstore-print` is transformed to `--print`; etc.