Switch to LLVM backend by default.
This will switch BAP to use LLVM as a default loader.
Also, this change the behavior of the `Image.create` function.
Previously, if backend wasn't specified it started the autoresolution
process. Now, it will default "llvm" also. But, if "auto" is provided
as a backend name, then the autoresolution will be started.