Final refining before release.
1. Refined `project` type

This PR is major refining of a project type, that is used in a plugin
system. First of all, some fields are deprecated, since we're going to
move away from the low-level representation to a more high-level
intermediate representation. Also, we're gradually moving away from the
incorrect abstractions of symbol (where the symbol must be contiguous in
space and every block belongs only to one symbol).

This PR also introduce a storage map, that can be used to exchange data
between plugins.

2. Renamed `bap-objdump` to `bap`

bap-objdump outgrew its name

3. Put Travis on a diet

Waiting for three hours is more than enough! Reduced the set of testing
files, but now we're testing on more platforms.

4. Removed examples folder

Since it is dead anyway.