Fix building documentation.
Releasing revealed a bug in documentation, that wasn't caught by CI,
since we're not compiling documentation on Travis. I've fixed this
bug in opam file, that is in the official repository, and bumped changes
back here. Also, I fixed the source of the error.

Since, that time we no longer build documentation as a part of
installation, since it takes too much time, very faulty (this is not a
first time), and worthless (the output is of very low quality). When
codoc will mature, we will return to this issue.