Fix possible double linking in bapbuild.
On a linking phase, we're excluding from the plugin all bap library
dependencies, but not the dependencies of the bap executable. As a
result we can link the same library twice, and if it has some kind
of state it can lead to undefined behavior, aka segmentation fault.

Since it is impossible to figure out what libraries are linked into
the bap binary in runtime, the only way it to hardcode this information
in the bapbuild itself. As a side effect, there is no need to define
`-pkg cmdliner,re.posix` to compile a plugin. In fact, all bap binary
dependencies are now automatically available.