Fix #181 and Block.dfs.
There was a bug in `Block.dfs` that results in that
dsf sequence of blocks was consumed after first run.

In the nature of an apology for this I also provided
an extra optional parameter `order` that allows you
to specify whether you would like visit blocks in
pre-order or post-order.

Also I fixed #181.