Introduce Intermediate Representation.
This PR adds long-awaited BAP IR. It is still work-in-progress (IR is
itself ready, but it is still far from being released). So I'm not going
to advertise the change right now. But everyone is welcome to review,
anyway.

Also, this PR moved stuff from `Tag` module to `Value`, and adds
`Value.Map` data structure.