Change Var module implementation.
This PR introduces a slight change in `var` representation.
Now variables can be versioned, i.e., a version number can be attached
to a variable to represent the same variable but in different time or
space. This is in particular useful for SSA form. See documentation in
Var module for more information.

Also small changes:

- Fix bug, that was introduced when I fixed #238. Due to the new bug,
  some blocks in IR became unreachable.

- BIL lifters now use sameness to compare variables. I.e., `CPU.is_*`
  works fine in SSA form.

- Added three new functions to Term module:
  - `change` for upsert operation
  - `find_exn` because sometimes it can be guaranteed
  - `enum` as a synonym to `to_sequence`

- added new option `output` as a synonym to `phoenix`.
  This will fix #241.