This PR tentatively introduces Graphlib library.
Graphlib extends OCamlGraph with a new modern and more rich interface.
It also provides:

- two new graph types
- depth_first_search algorithm
- dominators algorithm
- dominance frontier
- strongly connected components
- shortest path and reachability algorithms
- auxiliary data structures, like Rose Trees, Partition Sets, Frontiers,
  Paths, etc;
- functor that implements new interface on top of ocamlgraph modules
- functor that implements an old ocamlgraph interface on top of modules,
  that implements a newer [Graph] interface;
- pretty-printing for all graphs and auxiliary data structures;
- generic dot printing.

The library uses first class modules instead of functors, thus making
working with graph syntactically easier.

Note: this is a pre-release. I feel there still rough edges (and nodes)
left in the library. But everyone is welcome to play with it. Feedback
is very welcome.