Fix #235.
This is actually a workaround, since the real problem is still there:
```
 let high = Addr.of_int64  0xffffffffff601000L;;
 Addr.to_int64 high;;
- : int64 Core_kernel.Std.Or_error.t = Core_kernel.Result.Error (Z.Overflow)
```

The problem is that doesn't fit into OCaml's signed int64. As a
workaround, we can use `Addr.to_int64 (signed high)` in that case,
we will get a number with corresponding bit-pattern. For the purpose of
printing it is OK, but not in general.