Refined graph, arg term and abi interfaces.
Graph
-----

This PR introduces a mathematical model of a graph data structure.
Graph implementations are updated to conform to the model, so that
now they are much more consistent.

The difference between the identity of a node and its label is made
more clear. As a result, the node type of IR graph is made abstract.

Also, multigraphs (i.e., graphs, that allows more than one edge between
the same two nodes) are disallowed, since they do not fit into the
model, and the same relation can be represented by using suitable
labels.

Functor `Labeled` was added to the library. It will create a labeled
node.

Old interfaces that used ocamlgraph were removed and substituted with
newer and more clean, that use graphlib. Dfs function is removed, to
signal, that one should use graph interface instead.

Arg term
--------

Arg term is refined to be a some sort of definition. This is a part of
work of implementing argument recognition.

ABI interface
-------------

Abi interface is now updated to take sub term instead of lower level
`Symtab.fn`.

As a free bonus: sequences are made printable (not as printable) as
lists are, but shallow sequences of printables will be printed.