Move CI system to a container based infrastructure
This PR greatly simplifies the CI system. First we switch to a new OCaml
integration with Travis, that allows us to drop off our own system and
use `opam` file as a single point of truth. Second, it uses an
experimental branch of the integration system, that uses modern
container based infrastructure, that Travis enforces to use.

The pros, is that time to test is severely reduced, from 3 hours to 20
minutes. Also, it is expected that the new infrastructure is much more
stable. (At least, I hope that it wouldn't be as unstable as the legacy
one).

There're many cons also:
- no more checks for multiple compiler versions.
- no more excessive tests on our suite of test binaries.

This issues should be addressed at some point. Either by creating our
own CI locally, that is more powerful and doesn't have severe time and
memory restrictions. Or by still using travis, or some other online
services, but running this tests on a night-builds approach.