Add & organize serialization for BIL & BIR
This commit codifies a set of translation functions between
BIL and BIR for bap in such a way that single functions that
serialize all the way between each of the respective ocaml
vernacular types to strings and back. It is complete with
unit tests to demonstrate good working order and
functionality, all of which are passing. Lastly, this commit
reorganizes the extant bil code from the serialization library
and out of the main bap.bap_types sub-library itself into it's
own subfolder within bap.

Future directions after ocaml concurrency and other external
limitations are lifted (including core_kernel ceiling version,
and modular implicits), and of course ppx_protobuf_deriving is
duly noted.