Allow to introduce new tag values in plugins.
This PR allows to introduces new tags in plugins and pass values
of the specified tag between two different plugins.

Previously it wasn't possible, since a module that defines a tag
was evaluated per each plugin, and a failure exception was thrown
after the first evaluation, stating that the tag value is already
registered.

This PR fixes this in the following way. On a first registration
we store the provided module as well as freshly create tag value
in the info data structure. On consequent registration we look at
the type info table and if we find the same module as a user is
trying to register then we're returning a stored tag, otherwise
we bail out with an error. Of course, OCaml type system will not
allow us to extract a tag value from monomorphic table, so here
we use dark magic, and type cast it. It is unsound, but can be
considered safe, because we have at least two witnesses of type
equality. The first witness is that uuid of both types is equal
(but this is not enough, since a user can accidentally (by copy
pasting) provide the same uuid for two different types). The
second witness is that both types share the same comparison function.
It looks like, that the latter is enough to prove that the types,
are the same, since we're requiring to pass a module as a value,
and packing actually creates a record with functions. For example,
the following two modules will not be considered equal:

```ocaml
module X = struct
  type t = int
  let compare = compare

module Y = struct
  type t = float
  let compare = compare
end
```

A full example, showing how a user can pass a value of arbitrary
type between two plugins is provided below:

                               (* definition.ml *)

    open Core_kernel.Std
    open Bap.Std

    let note = Value.Tag.register
        ~name:"note"
        ~uuid:"e4562d18-3ac5-4070-bc12-b99b9856dfce"
        (module String)

                              (* definition.mli *)

    open Bap.Std

    val note : string tag

                                 (* pass1.ml   *)

    open Bap.Std
    open Definition

    let main proj =
      Project.set proj note "hello"

    let () =
      Project.register_pass "pass1" main

                                 (* pass2.ml   *)

    open Bap.Std
    open Core_kernel.Std
    open Definition

    let main proj =
      match Project.get proj note with
      | Some note -> printf "Note: %s\n" note
      | None -> printf "No notes\n"

    let () = Project.register_pass' "pass2" main

compile with
```
bapbuild pass1.plugin && bapbuild pass2.plugin && bap /bin/true -lpass1 -lpass2
```

expected output:

```
Note: hello
```