Fixed IR lifting procedure
Several corner cases (that are likely occur only when we're lifting some
junk from data) are handled.

The cpuexn instruction lifting is completely rewritten, as previously it
was incorrect. The return label will target to a trampoline block, that
will either a terminal block, if no code follows the cpuexn, or will
redirect to a fall-through block. This actually raises a question,
whether we should make a return label optional in the interrupt term.

Added check that a call is not conditional, to handle conditional
calls. The check relies on a BIL representation, rather on a backend
kind mechanism, to ensure that there is no controversy between the
recursive descent disassembler and IR lifter.