Allow to pass options to LLVM backend.
Options can be passed to the backend using environment variable
`BAP_LLVM_OPTIONS`. To get the option list use: BAP_LLVM_OPTIONS="-help"
Currently there're only:
```
OPTIONS:
  -help                            - Display available options (-help-hidden for more)
  -mc-x86-disable-arith-relaxation - Disable relaxation of arithmetic instruction for X86
  -version                         - Display the version of this program
  -x86-asm-syntax                  - Choose style of code to emit from X86 backend:
    =att                           -   Emit AT&T-style assembly
    =intel                         -   Emit Intel-style assembly
```

With the most interesting option `-x86-asm-syntax`, that allows to
switch between `att` and `intel` syntax. (Actually this is the main
reason to add this mechanism, as it was requested by many of our users,
to provide output in the intel syntax).

The default syntax is still att.