Add bap data framework and refine variables
(sorry but it was impossible to split them into two different commits)

Bap Data Framework
==================

The purpose of the framework is to give a user more control on data
serialization and printing, in particular to allow a user to add new
serialization formats dynamically via plugin system. It also reduces
the amount of boilerplate code that should be written for any IO.

The PR adds a new interface `Data` that captures an idea of inductively
defined data, that has a representation, that can be serialized and
deserialized. In other words, data is something that has a real physical
representation in the program, unlike co-data that actually exists
somewhere outside of the program.

A Regular interface now include Data interface, so all Regular types (a
majority of data types in BAP are now implement the Data interface).

Data interface comes with lots of useful functions for serialization and
IO. This functions accepts optional arguments `?ver` and `?fmt` that
allows a user to switch serialization format on the fly. For example:
`to_bytes ~fmt:"sexp: var` and `to_bytes ~fmt:"bin" var` will serialize
a variable `var` into a sequence of bytes using `sexp` format or
`bin` (for binprot) formats correspondingly. Default writer, reader and
printers can be set permanently or temporary for each particular data
types.

This PR also ships two generic data types called `writer` and `reader`.
This data types represent `Read` and `Write` type classes. To make an
instance of a type class, it is sufficient to provide a minimum
implementation and everything else will be derived. This allows to
easily implement new formats without caring to much about performance
and boilerplate. On the other hand, such approach allows to fine tune
the implementation for performance, by providing custom implementation
for methods.

Variable Refinement
===================

Bap Variable data type got refined. The ugly `tmp` flag with unspecified
semantics is no longer there. Instead we have two orthogonal flags that
should be passed to a variable constructor:

[is_virtual] to designate that variable corresponds to a purely virtual
data, that has no real storage in the program. The opposite - real
variable, represents a variable that is either a register or memory cell
or memory storage. Virtual variables can be introduced by lifters and
analyses to represent intermediate results of calculations.

[fresh] if set to [true] will generate a random salt, using state
statically stored inside the module and mix it to a variable name.