split each library into a separate oasis section
* wrote automation that glues everything together
* fixed bap-fsi-benchmark
* brute-forced uninstallation

Now, every feature that is provided by the bap repository is described
in a separate file in the oasis folder. The configure scripts is
responsible to glue them together based on the `--enable-<feature>`
options. There is a `--enable-everything` flag, that is useful for
testing and development. The default way of configuring from the source
repository would be:
```
  ./configure --enable-everything --prefix `opam config var prefix`
```