robust plugin loader
This PR will make a plugin loader more robust to a malicious user. It
will not fail if a plugin is not a plugin, or if it can't be load. It
will print a warning and continue. In a verbose mode it will print an
extended error message.

This will fix #351 and partially resolve #353.