Delayed plugin initialization
Read packages written in core now delayed until the first load.
Previously, it was initialized at the library link time, that was ok
when we compiled from a trunk, but after we split the libraries, it
stated to lead to segfaults, since only packages written in the bap-std
package were taken into account.