support GNU attributes
The main feature of this PR is the support of the GNU attributes by the
api plugin. The GNU attributes are mapped to corresponding Term
attributes. See `Sub` module for a detailed description. The attributes
are added from libc headers. So far we understand the following GNU
attributes:
  - const
  - pure
  - malloc
  - noreturn
  - returns_twice
  - nothrow
  - warn_unused_result
  - alloc_size
  - format

Much more Term attributes were added, for the later use.

Also, this PR adds a `callsites` plugins, that will annotate callsites
with uses and definition of called subroutine. This is useful, in
particular, for data flow analysis based on taint analysis, and is
required for Saluki (in fact this is taken from Saluki). The `taint`
pass no has `callsites` in its dependencies. Here is a usual excerpt
from the man page for the callsites plugin (--callsites-help):

       This pass will inject artifical definitions of a subroutine
       arguments at call sites. Consider function malloc that has the
       following declaration in BIR:

             sub malloc(malloc_size, malloc_result)
             malloc_size :: in u32 = R0
             malloc_result :: out u32 = R0

       This plugin will add two definitions, one just before the call to
       the malloc:

             ...
             000001c3: malloc_size := R0
             0000015b: call @malloc with return %0000015c

       And prepend another to the block to which malloc will return:

             0000015c:
             000001c4: R0 := malloc_result
             ...