better taint propagation
This PR includes several improvements and bug fixes spread across
the code base, united by the main goal - improve taint propagation
framework. It is also found a necessary base for the upcoming movement
of Saluki to the new framework.

Main contribution
-----------------

It was observed, that a concretization to a single value will lead to
unwanted aliasing between arbitrary pointers. Consider the following
example:
```
000001dc: R0 := strcpy_dest
0000016d: R3 := mem[R11 - 0x34:32, el]:u32
```

Here `*strcpy_dest` and `*(R11-0x34)` a both unknown and will have the
same value. As a result `R0` and `R3` will point to the same memory
cell.

Note - this problem manifests itself with tainted pointers more often,
than with tainted immediates, mostly because for a tainted immediate
value we distinguish between different computations that results to the
same value. We cannot do this with pointers, as we will loose
precision.

This PR adds four more options to the concretizer class, three of them
are propagated to the command line interface:

- `lookup` function (not available on command line) - allows a user to
  provide concrete values for unknown variables. It is useful, for
  example, to resolve SP to a value close to the `stack_start`, that
  will reduce the probability of unwanted aliasing between heap memory
  and stack memory. The `propagate-taint` plugin is setting SP to some
  high constant be default.

- `reg_policy` - a policy for concertizing unknown register
  (defaults to `random`)

- `mem_policy` - a policy for concertizing memory loads
  (defaults to `random`)

- `random_seed` - a seed for RNG (useful to get predictable results)

Auxiliary stuff
---------------

- Added printf and scanf family of functions to the api/posix
- fixed parsing attributes, so that format attribute is now can be read
- added enter/leave term methods to the Term.visitor
- fixed few bugs in callsites plugin