warn-unused pass
This PR adds a simple analysis that outputs a warning for each function
that is marked with a `warn_unused_result` attribute and whose result
value is dead. Unlike an unchecked result, that we use in Saluki, this
plugin has a higher threshold, it will warn if the result is dead, it
doesn't require it to be checked.

Also this PR fixes a bug in the taint propagation pass.

Here comes the printout of the man page (bap --warn-unused-help):

```man
WARN-UNUSED(1)                Warn-unused Manual                WARN-UNUSED(1)

NAME
       warn-unused - warn about unused result of certain functions

SYNOPSIS
           --warn-unused
           --warn-unused-taint
           --warn-unused-print
           --warn-unused-mark

DESCRIPTION
       If a subroutine has GNU attribute warn_unused_result and its result is
       not used, then print a warning message.

OPTIONS
       --help[=FMT] (default=pager)
           Show this help in format FMT (pager, plain or groff).

       --version
           Show version information.

PASSES
       --warn-unused-taint
           Taint all values defined by functions that are marked with
           warn_unused_result attribute. Will run callsites as a dependency.

       --warn-unused-print
           Print all calls that weren't checked.

       --warn-unused-mark
           Mark all unchecked calls with Term.dead attribute

       --warn-unused
           Same as --warn-unused-taint --propagate-taint --warn-unused-print

Warn-unused 0.1                                                 WARN-UNUSED(1)
```