more accurate taint propagation
This PR is a bunch of fixes and reification, that in sum
leads to a better (more accurate) taint propagation and
analysis.

Move some attribute to arg module
---------------------------------

Some GNU attributes, that are defined on a function level, are actually
should be assigned to attributes. This attributes, are often use some
number to point on a positional parameter, like `format(printf,1)` show
that first parameter is a printf format string. Now, we attach such
arguments not to the function, but to the argument of subroutine. The
following attributes are affected:

  - alloc_size
  - format
  - nonnull (new)
  - restrict (new - in fact it is type qualifier)
  - warn_unused (previously known as warn_unused_result)

Also three new attributes are added:

  - precondition
  - invariant
  - postcondition

They can be attached to a term. An expression, associated with attribute
is a boolean `BIL` expression, that must be true at corresponding
positions.

A change in a taint introduction mechanism
------------------------------------

Now, if a term marked as tainted, then that means that all
values, produced in this term are for some reason tainted.

For example, previously if we had:
```
.tainted_reg %1
R0 := R1
```

Then we introduced taint into `R0`, leaving `R1` intact.

Now, in this definition we will mark `R1` as tainted, and the taint will
propagate into `R0` as a side-effect.

With this new approach it is possible now to fix callsites plugin. Now
all synthetic terms, that are used to fix dataflow from and out of the
function will have arguments on the left hand side. This is, first of
all, more intuitive, and, most important, will fix an issue, when an
inout argument killed return value. Consider the following example,
previously we put the following definitions around the call to `strcpy`

```
  strcpy_dst := R0
  strcpy_src := R0
  call @strcpy return %1
1:
  R0 := strcpy_result
  R0 := strcpy_dst
```

Beside that it is incorrect at all, it also has a nasty feature, that
the last definition kills the previous. But after this PR, the
following definitions are inserted:

```
  strcpy_dst := R0
  strcpy_src := R0
  call @strcpy return %1
1:
  strcpy_result := R0
  strcpy_dst := R0
```

Proper taint propagation from/to memory
---------------------------------------

- load to register will kill all taints stored in the register
  and propagate taints associated with the address
- store will kill all taints associated with address and substitute
  them with taints associated with a result stored in the register.

Note: previously it was a union of taints, that resulted in
overtainting.

Fix API parsing
---------------

- recognize restricted pointers
- recognize arrays
- make intent non-optional