This PR introduces traces to core bap.
The PR removes `bap-tracedump` utility, and reintroduces it
as a bap plugin. The plugin can load and dump traces. A command

```
bap --trace-dump file
```

is equivalent to the old
```
bap-tracedump file
```

The main feature, is the ability to load traces, so that they can be
later accessed during the analysis. See corresponding documentation in
module Bap_traces.Std.

Detailed changes
================

1. The bap-traces library is updated and partially re-implemented,
   now it has a constant memory footprint. Now it has a flavor of
   an input channel, e.g., operation `events` is renamed into
   `read_events.

2. Refactored meta information, made it more usable

3. Introduced event streams and futures to the BAP
   (please, don't rush to use it, it may change in the nearest future)

See also PR to bap-frames and qemu, that reflects the changes to the
main bap repo.